requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.26.0
Pillow==10.0.0
scikit-learn==1.6.1
setuptools>=65.5.0
wheel>=0.38.0 